change db volumes to float

+++ src/pyblu/_entities.py
@@ -27,14 +27,14 @@ class Status:
 
     volume: int
     """Volume level with a range of 0-100"""
-    volume_db: int
+    volume_db: float
     """Volume level in dB"""
 
     mute: bool
     """Mute status"""
     mute_volume: int | None
     """If the player is muted, then this is the unmuted volume level. Absent if the player is not muted."""
-    mute_volume_db: int | None
+    mute_volume_db: float | None
     """If the player is muted, then this is the unmuted volume level in dB. Absent if the player is not muted."""
 
     seconds: int
